Add health check route
